Simplify test file loading
Loading files for tests currently involves a lot of manually-constructed
relative requires, as well as a Dir.chdir in the tests/all_tests.rb file
which adjusts where requires need to be relative from.

Apart from some boilerplate, the main effect of this is that it's hard
to run a single test file from the command line, as you need to edit the
all_tests.rb file.

This makes things a bit simpler by:

* Adding the tests directory to the load path for the `rake test` task
* Removing the chdir from all_tests.rb to avoid a bunch of path
  complexity
* Replacing all relative requires with direct ones.

You can now run a single test file from the root directory of the
project in a fairly standard way:

    $ ruby -Ilib:tests tests/units/test_object.rb

Signed-off-by: Simon Coffey <simon.coffey@futurelearn.com>