Remove needless bare repo clone
The 'working' fixture repo has a remote configured pointing at the
'working.git' bare fixture repo. This assertion appears to be checking
that you can archive, via the main repo object, something in the remote.

There doesn't appear to be any reason you need to clone the remote repo
to make this assertion, so this removes that step.