Introduce in_temp_repo helper to avoid unnecessary clones
In a number of tests we use the set_file_paths helper, which creates a
temporary clone of the 'working' fixture repo, and then create
subsequent clones of the 'working.git' bare fixture repo.

This duplicates work unnecessarily so I'd like to eliminate it. This
introduces a new helper, in_temp_repo, that both clones a named fixture
repo and moves to its directory, which can then be used for whatever
operations a given test needs.